refactor: 明确 Dashboard 数据服务职责
将 DashboardModel 迁移为 feature application 层的 DashboardDataService，并分离展示用的 DashboardUpcomingEntryData；保持 API 请求、缓存预热和学期解析行为不变。
