+++ lib/features/dashboard/application/dashboard_data_service.dart
@@ -1,26 +1,14 @@
+import 'package:onetj/app/di/dependencies.dart';
+import 'package:onetj/models/course_schedule_data.dart';
+import 'package:onetj/models/school_calendar_data.dart';
+import 'package:onetj/models/student_info_data.dart';
+import 'package:onetj/repo/course_schedule_repository.dart';
+import 'package:onetj/repo/student_info_repository.dart';
+import 'package:onetj/services/term_key_resolver.dart';
 import 'package:onetj/services/tongji.dart';
 
-import 'package:onetj/app/di/dependencies.dart';
-import 'package:onetj/models/student_info_data.dart';
-import 'package:onetj/repo/student_info_repository.dart';
-import 'package:onetj/models/school_calendar_data.dart';
-import 'package:onetj/models/course_schedule_data.dart';
-import 'package:onetj/repo/course_schedule_repository.dart';
-import 'package:onetj/models/timetable_index.dart';
-import 'package:onetj/services/term_key_resolver.dart';
-
-class DashboardUpcomingEntryData {
-  const DashboardUpcomingEntryData({
-    required this.entry,
-    required this.isOngoing,
-  });
-
-  final TimetableEntry entry;
-  final bool isOngoing;
-}
-
-class DashboardModel {
-  DashboardModel({
+class DashboardDataService {
+  DashboardDataService({
     TongjiApi? api,
     TermKeyResolver? termKeyResolver,
   })  : _api = api ?? TongjiApi(),
