+++ lib/features/dashboard/models/dashboard_upcoming_entry_data.dart
@@ -1,0 +1,11 @@
+import 'package:onetj/models/timetable_index.dart';
+
+class DashboardUpcomingEntryData {
+  const DashboardUpcomingEntryData({
+    required this.entry,
+    required this.isOngoing,
+  });
+
+  final TimetableEntry entry;
+  final bool isOngoing;
+}
