+++ lib/features/dashboard/views/dashboard_view.dart
@@ -10,7 +10,7 @@ import 'package:onetj/app/theme/theme_change_notifier.dart';
 import 'package:onetj/features/app_update/app_update_flow_coordinator.dart';
 import 'package:onetj/features/dashboard/view_models/dashboard_view_model.dart';
 import 'package:onetj/features/dashboard/views/widgets/function_grid.dart';
-import 'package:onetj/features/dashboard/models/dashboard_model.dart';
+import 'package:onetj/features/dashboard/models/dashboard_upcoming_entry_data.dart';
 import 'package:onetj/models/app_update_info.dart';
 import 'package:onetj/models/dashboard_upcoming_mode.dart';
 import 'package:onetj/app/presentation/ui_event.dart';
